Made new branch on DDSA-ROS called linh
New branch contains changes from DDSA_ROS repository with 1 new file rover.msg (custom message type used inside mobility.cpp).

Changes to mobility.cpp:

1) Created a timer called roverCountTimeer triggers for 10 secs wait before the stimulation runs to register the numberOfRovers to generate the rovers pattern inside the ddsa_controler.

2) Created a subscriber inside mobility.cpp that subscribes to /numberOfRovers and calls roverCountHandler. roverCountHandler takes in a rover type message from the mobility package that contains two data types bool (moveToNest) and string (roverName). roverCountHandler push and pop string items inside the globally declared priority_Queue called roversQ. push and pop is triggered by the moveToNest value. moveToNest is initialized as false and triggered true inside. Also it counts the number of rovers running in the stimulation called numberOfRovers.

3) Inside stateMachineState movements are made when a new goalLocation is set. There are 4 cases manage the rovers movements.

A) Check if the moveToNest is false and that the rovers is at most 0.1 meters way form the center and that the current rover is first in the queue. Then we reset its goalLocation to be at the center.

B) Else if moveToNest is false and that rover is within 0.1 radius of the center. Then set moveToNest to true and roverUpdate ojbecjt will publish to the numberOfrovers topic to all the rovers who subscribe to it that moveToNest and the rover that moved to the center status. At this point the roverCountHandler will pop that rover that moved to the nest of the queue.

C) Else if moveToNest is true then the rovers will get the next goalLocation in the pattern. This case will only happen once the rover has already been pop out of the priority queue.

D) Else the current rover is not next in the queue so it needs to wait for the current queued rover. If the rover is within 1.5 meters of the nest then the rover move a
 radialDistance and face way from the center. Else if rover is far enough it can stop where it is currently at.

4) Created a bool called setSpiralLocation that prevents the rovers from collecting additional tags before it reaches the nest. setSpiralLocation is used inside targetHandler.

5) Created a bool called setObstacleAvoidance (currently not working) that will eventually correct the rovers goalLocation to be to their original goal before the collision happpened.