added GUI elements for the global offset
A checkbox has been added in the GUI so that the user can toggle
how the map frame is displayed.

When toggled off, the map will display as it always has with the
rovers paths all drawn from point (0,0).

When toggled on, the rover paths will be offset by their starting
positions.

This distinction is important so that the DDSA spirals are displayed
properly in the GUI map frame.