updated MapData w/ comments verifying compliance with physical rovers
The physical rovers were able to run without crashing or errors with
this feature branch. In addition to that, further research was done
to ensure that the code that sets the global offsets for the map frame
will not break anything else in the code or simulation.

The C++ std library containers used to keep track of the offsets will
create appropriate defaults when the global offsets are not initialised
(as will be the case when we use physical rovers). Please refer to
comments added to the appropriate code for details.

Future updates to this functionality could be used to pass a global
offset representing the rover's physical starting position in the real
world. For now, clicking on the check box for this feature in the GUI
has no effect, which is intended and correct. This is currently a
simulated rover only feature.